fix/github: return useful value instead of map
Returning `actor` map in detections is generally a bad idea and will
force some errors.
Better to return the properties we want. It's also possible to use
`Fields` to surface other data points we want to.

But in this case, `actor` is put in the `results` array and when it's
processing it, it does not look for the whole object.